# openapi-output/@typespec/openapi3/openapi.yaml
openapi: 3.0.0
info:
  title: (title)
  version: 0.0.0
...
paths:
...
  /api/bookings:
    post:
      operationId: GuestApi_createBooking
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - Guest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                eventTypeId:
                  type: string
                slotTime:
                  type: string
                guestName:
                  type: string
                guestEmail:
                  type: string
              required:
                - eventTypeId
                - slotTime
                - guestName
                - guestEmail
...
  /api/event-types/{eventTypeId}/slots:
    get:
      operationId: GuestApi_getSlots
      parameters:
        - name: eventTypeId
          in: path
          required: true
          schema:
            type: string
        - name: from
          in: query
          required: true
          schema:
            type: string
          explode: false
        - name: to
          in: query
          required: true
          schema:
            type: string
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Slot'
      tags:
        - Guest
components:
  schemas:
...
    Booking:
      type: object
      required:
        - id
        - eventTypeId
        - slotTime
        - guestName
        - guestEmail
        - createdAt
      properties:
        id:
          type: string
        eventTypeId:
          type: string
        slotTime:
          type: string
        guestName:
          type: string
        guestEmail:
          type: string
        createdAt:
          type: string
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
...
    Slot:
      type: object
      required:
        - startTime
        - eventTypeId
        - status
      properties:
        startTime:
          type: string
        eventTypeId:
          type: string
        status:
          $ref: '#/components/schemas/SlotStatus'
    SlotStatus:
      type: string
      enum:
        - available
        - booked